fixed displaying of spectrum and structure on FwdRecords.jsp
instead of just forwarding the request to

Dispatcher.jsp?type=disp&id" + id + "&site=" + val[1];

the request will be redirected to

FwdRecord.jsp?...&dsn=dbname

if the dsn parameter was missing in the url of the request. Then the request will be forwarded to Dispatcher.jsp from there.

Therefore, the dsn parameter is kept in the browser url and the java script componenents can use it to load data for spectrum and structure display.